ENG-7409 Move host app interaction to an interface. See PR descriptrion for details. inital commit
