+++ NeuroID/src/test/java/com/neuroid/tracker/events/NIDIdentityAllViewsTest.kt
@@ -10,7 +10,7 @@ import android.widget.RatingBar
 import android.widget.SeekBar
 import android.widget.ToggleButton
 import androidx.appcompat.widget.SwitchCompat
-import com.neuroid.tracker.NeuroID
+import com.neuroid.tracker.NeuroIDImpl
 import com.neuroid.tracker.utils.NIDLogWrapper
 import io.mockk.every
 import io.mockk.just
@@ -219,8 +219,8 @@ class NIDIdentityAllViewsTest {
         }
     }
 
-    private fun mockNeuroID(): NeuroID {
-        val nidMock = mockk<NeuroID>()
+    private fun mockNeuroID(): NeuroIDImpl {
+        val nidMock = mockk<NeuroIDImpl>()
         every {
             nidMock.captureEvent(
                 any(),
