+++ app/src/main/java/com/sample/neuroid/us/activities/DynamicActivity.kt
@@ -6,8 +6,8 @@ import android.view.ViewGroup
 import android.widget.EditText
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
-import com.neuroid.tracker.NeuroID
-import com.neuroid.tracker.utils.NIDLog
+import com.neuroid.tracker.NeuroIDImpl
+import com.sample.neuroid.us.MyApplicationDemo
 import com.sample.neuroid.us.R
 import com.sample.neuroid.us.databinding.NidActivityDynamicBinding
 
@@ -41,6 +41,6 @@ class DynamicActivity : AppCompatActivity() {
 
     private fun addNewViewWithRegisterTarget() {
         // re-register targets after we add an EditText to the layout.
-        NeuroID.getInstance()?.registerPageTargets(this)
+        NeuroIDImpl.getInstance()?.registerPageTargets(this)
     }
 }
